Create cython point_accretion module
The PointAccretor baseclass handles the accretion algorithm, while
subclasses are responsible for handling the method `is_bin_full()` to
test if the bin is fully formed.

Currently I've implemented the EqualMassAccretor that attempts to build
bins with a minimum target mass.

The code uses scipy cKDTree to do nearest neighbour searches.