Optimize PointAccretion find_closest_unbinned
Before I would re-build the list of points that are unbinned and run a
KD Tree search against that to find the point closest to the node. This
required a lot of instantiations of cKDTree and array rebuilding!

Now I build a single KD tree of all points, and allow myself to search
for multiple points close to a node. I select the first unbinned point
from a node. This seems to be about 4x faster than before.