Create pixel_accretion module
The pixel_accretion module is intended to be equivalent to
point_accretion, but for explicity working with pixel grids.
The PixelAccretion class accretes only pixels on the edge of the current
bin. Doing so allows users to build non-circular regions to capture, say,
iso-colour and iso-photal pixel groups.

Users are expected to subclass PixelAccretion to implement
science-driven binning rules. For example, IsoIntensityAccretor
implements binning to ensure regions have intensity standard deviations
within a user-defined threshold.

Added a test in tests/ that builds a 16x16 pixel image in four distinct
intensity groups.